Real django 1.8 support!
The new test runner of django 1.6 (that had to be
used for django 1.8) did not automatically discover
tests from test/__init__py, so the tests *were not*
running for django 1.8! To fix that, all test modules
were renamed to start with test_* so now they are
discovered.

Unfortuanately, this leaded to a bunch of tests failing
in django 1.8 because:

- the fields attribute is required in a CBV please see
https://docs.djangoproject.com/en/1.8/topics/class-based-views/generic-editing/#model-forms
(fixed in module tests.class_decorator_classes by adding all fields
of the User class)
- the get_form doesn't require the form_class parameter anymore,
please see https://docs.djangoproject.com/en/1.8/releases/1.8/#generic-views
(fixed in module class_decorator by checking if version >= 1.8 and
defining the new_get_form method accordingly)